requirements.txt
Appium-Python-Client
selenium
pytest
python-dotenv
